Upload & upload with programmer function properly
When using the upload commands, call compile with the appropiate --upload flags, rather than using the raw upload command.  This way, the code will be compiled and uploaded, matching the behavior of the non-cli path.

Also fix the specification of '--build-path' for uploads.

It may be worth adding an 'upload only' functionality in the future.